user can add multiple product and remove from cart

+++ src/app/components/cards/CartItem.jsx
@@ -4,6 +4,7 @@ import Image from "next/image";
 import { FaPlus, FaMinus, FaTrash } from "react-icons/fa";
 import Swal from "sweetalert2";
 import toast from "react-hot-toast";
+import { deleteItemsFromCart } from "@/actions/server/cart";
 
 const CartItem = ({ item, onQuantityChange, onRemove }) => {
   const { _id, image, title, quantity, price } = item;
@@ -25,17 +26,21 @@ const CartItem = ({ item, onQuantityChange, onRemove }) => {
       icon: "warning",
       showCancelButton: true,
       confirmButtonText: "Yes, remove it",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const result = await deleteItemsFromCart(_id);
+        if (result.success) {
+          Swal.fire("Deleted", "This item successfuly removed from cart", "success");
+          onRemove(_id); 
+        } else {
+          Swal.fire("Error", "Failed to delete", "error");
+        }
+      }
     });
-
-    if (result.isConfirmed) {
-      onRemove(_id);
-      toast.success("Item removed from cart");
-    }
   };
 
   return (
     <div className="flex items-center gap-4 p-4 border rounded-xl bg-base-100 shadow-sm">
-
       {/* Image */}
       <div className="w-20 h-20 overflow-hidden rounded-lg">
         <Image
@@ -55,19 +60,13 @@ const CartItem = ({ item, onQuantityChange, onRemove }) => {
 
       {/* Quantity Control */}
       <div className="flex items-center gap-2">
-        <button
-          onClick={handleDecrease}
-          className="btn btn-sm btn-circle"
-        >
+        <button onClick={handleDecrease} className="btn btn-sm btn-circle">
           <FaMinus />
         </button>
 
         <span className="px-2 font-semibold">{quantity}</span>
 
-        <button
-          onClick={handleIncrease}
-          className="btn btn-sm btn-circle"
-        >
+        <button onClick={handleIncrease} className="btn btn-sm btn-circle">
           <FaPlus />
         </button>
       </div>
